@@ -9,7 +9,7 @@ import android.graphics.Typeface;
 import android.util.AttributeSet;
 
 
-public class DancingTextView extends android.support.v7.widget.AppCompatTextView {
+public class DancingTextView extends androidx.appcompat.widget.AppCompatTextView {
     public DancingTextView(Context context, AttributeSet attrs) {
         super(context, attrs);
         this.setTypeface(Typeface.createFromAsset(context.getAssets(), "fonts/dancing_font.otf"));
