@@ -9,7 +9,7 @@ import android.graphics.Typeface;
 import android.util.AttributeSet;
 
 
-public class VioletEyeTextView extends android.support.v7.widget.AppCompatTextView {
+public class VioletEyeTextView extends androidx.appcompat.widget.AppCompatTextView {
     public VioletEyeTextView(Context context, AttributeSet attrs) {
         super(context, attrs);
         this.setTypeface(Typeface.createFromAsset(context.getAssets(), "fonts/violet_eyes.otf"));
