@@ -11,14 +11,14 @@ import android.graphics.BitmapFactory
 import android.os.AsyncTask
 import android.os.Bundle
 import android.provider.MediaStore
-import android.support.design.widget.FloatingActionButton
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.*
-import android.support.v7.widget.Toolbar
-import android.support.v7.widget.helper.ItemTouchHelper
+import com.google.android.material.floatingactionbutton.FloatingActionButton
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.*
+import androidx.appcompat.widget.Toolbar
 import android.text.TextUtils.isEmpty
 import android.view.*
 import android.widget.*
+import androidx.recyclerview.widget.*
 import com.github.clans.fab.FloatingActionMenu
 import com.tanka.accessories.todoroom.R
 import com.tanka.accessories.todoroom.data.model.Note
