@@ -3,7 +3,7 @@ package com.tanka.accessories.todoroom.views
 import android.content.Intent
 import android.os.Build
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.view.View
 import android.widget.FrameLayout
 import android.widget.ImageView
