@@ -1,10 +1,10 @@
 package com.tanka.accessories.todoroom.data.room;
 
-import android.arch.persistence.db.SupportSQLiteDatabase;
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.Room;
-import android.arch.persistence.room.RoomDatabase;
-import android.arch.persistence.room.migration.Migration;
+import androidx.sqlite.db.SupportSQLiteDatabase;
+import androidx.room.Database;
+import androidx.room.Room;
+import androidx.room.RoomDatabase;
+import androidx.room.migration.Migration;
 import android.content.Context;
 
 import com.tanka.accessories.todoroom.data.model.Note;
@@ -13,7 +13,7 @@ import com.tanka.accessories.todoroom.data.model.Note;
  * Created by access-tanka on 11/16/17.
  */
 
-@Database(entities = {Note.class}, version = 2)
+@Database(entities = {Note.class}, version = 1, exportSchema = false)
 public abstract class AppDataBase extends RoomDatabase {
 
     public static final String DB_NAME = "note_db";
