@@ -1,6 +1,6 @@
 package com.tanka.accessories.todoroom.utility;
 
-import android.arch.persistence.room.TypeConverter;
+import androidx.room.TypeConverter;
 
 import java.util.Date;
 
